--- cmd/modular-indexer-light/cmd.go
@@ -1,1 +1,0 @@
-package main
