+++ internal/wallet/types.go
